+++ otx-pool/Cargo.toml
@@ -30,4 +30,3 @@ tokio = { version = "1", features = ["sync", "process"] }
 otx-format = { path = "../otx-format" }
 otx-plugin-protocol = { path = "../otx-plugin-protocol" }
 otx-pool-config = { path = "../util/config" }
-utils = { path = "../utils" }
