+++ otx-sdk/Cargo.toml
@@ -20,4 +20,3 @@ serde_json = "1.0"
 otx-format = { path = "../otx-format" }
 otx-plugin-protocol = { path = "../otx-plugin-protocol" }
 otx-pool-config = { path = "../util/config" }
-utils = { path = "../utils" }
