+++ plugins-built-in/atomic-swap/Cargo.toml
@@ -8,7 +8,6 @@ edition = "2021"
 [dependencies]
 anyhow = "1.0"
 ckb-jsonrpc-types = "0.105"
-ckb-stop-handler = "0.105"
 ckb-sdk = "2.3.0"
 ckb-types = "0.105"
 dashmap = "5.4.0"
