+++ plugins-built-in/dust-collector/Cargo.toml
@@ -7,15 +7,13 @@ edition = "2021"
 
 [dependencies]
 anyhow = "1.0"
+ckb-jsonrpc-types = "0.105"
 ckb-sdk = "=2.3.0"
-ckb-stop-handler = "0.105"
 ckb-types = "0.105"
 dashmap = "5.4.0"
-derive_more = "0.99"
 log = "0.4.17"
 
 otx-format = { path = "../../otx-format" }
 otx-plugin-protocol = { path = "../../otx-plugin-protocol" }
 otx-pool-config = { path = "../../util/config" }
 otx-sdk = { path = "../../otx-sdk" }
-utils = { path = "../../utils" }
