+++ plugins-built-in/dust-collector/src/lib.rs
@@ -1,3 +1,7 @@
+mod helper;
+
+use helper::{OutputAmount, TxBuilder};
+
 use otx_format::jsonrpc_types::OpenTransaction;
 use otx_plugin_protocol::{
     HostServiceHandler, MessageFromHost, MessageFromPlugin, Plugin, PluginInfo, PluginMeta,
@@ -5,7 +9,6 @@ use otx_plugin_protocol::{
 use otx_pool_config::built_in_plugins::DustCollectorConfig;
 use otx_pool_config::{CkbConfig, ScriptConfig};
 use otx_sdk::build_tx::OtxBuilder;
-use utils::aggregator::{OtxAggregator, OutputAmount};
 
 use anyhow::{anyhow, Result};
 use ckb_sdk::rpc::ckb_indexer::{Order, ScriptType, SearchKey};
@@ -197,7 +200,7 @@ impl Plugin for DustCollector {
             capacity: HumanCapacity::from(output_capacity),
             udt_amount: None,
         };
-        let aggregator = OtxAggregator::new(
+        let aggregator = TxBuilder::new(
             self.context.ckb_config.clone(),
             self.context.script_config.clone(),
         );
