+++ plugins-built-in/signer/Cargo.toml
@@ -7,14 +7,14 @@ edition = "2021"
 
 [dependencies]
 anyhow = "1.0"
+ckb-jsonrpc-types = "0.105"
 ckb-sdk = "=2.3.0"
-ckb-stop-handler = "0.105"
 ckb-types = "0.105"
 dashmap = "5.4.0"
-derive_more = "0.99"
 log = "0.4.17"
+secp256k1 = { version = "0.24", features = ["recovery"] }
 
 otx-format = { path = "../../otx-format" }
 otx-plugin-protocol = { path = "../../otx-plugin-protocol" }
 otx-pool-config = { path = "../../util/config" }
-utils = { path = "../../utils" }
+otx-sdk = { path = "../../otx-sdk" }
