--- utils/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "utils"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-anyhow = "1.0"
-ckb-crypto = "0.105"
-ckb-hash = "0.105"
-ckb-jsonrpc-types = "0.105"
-ckb-sdk = "=2.3.0"
-ckb-types = "0.105"
-faster-hex = "0.6"
-log = "0.4"
-once_cell = "1.10"
-rand = "0.8"
-secp256k1 = { version = "0.24", features = ["recovery"] }
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-toml = "0.5"
-
-otx-format = { path = "../otx-format" }
-otx-pool-config = { path = "../util/config" }
