--- utils/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod aggregator;
-pub mod build_tx;
